+++ tests/multigrid/test_p_multigrid.py
@@ -116,7 +116,7 @@ def test_prolong_low_order_to_restricted(tp_mesh, tp_family, variant):
         P = prolongation_matrix_matfree(uc, v).getPythonContext()
         P._prolong()
 
-    assert norm(ui + uf - uc, "L2") < 2E-14
+    assert norm(ui + uf - uc, "L2") < 1E-13
 
 
 @pytest.fixture(params=["triangles", "quadrilaterals"], scope="module")
